Better Scritps
As if my screen profusely burning my Retinas with an unescapable swamp
of green was not a good enough reason to remind me that we're using the
masterfully designed bit of software known as Robot Builder, it also
likes to elegantly remind us that we've been graced with its software
design by swamping our collective computers with trails of residual
temporary files that aren't actually all that temporary. For those who
are simply overwhelmed by these files, I have humbly included scripts to
Delete them.

(These files are gitignored, but in the grand scheme of things are still
incredibly obnoxious)