Base Class for stall detection on encoders & jags
Add Stall Detection to a jag or encoder in a subsystem. Simply extend
off StallableJag or StallableEncoder and implement an abstract method
that returns a pointer to the encoder or jag that you want to watch.
From your subsystem call FillStallArray() to store voltage data. The
idea behind this is that we don't have to reuse this code a million
times